Flytt getAntallUttaksdagerIVinduRundtFødsel til periodeUtils
Funksjonen var duplisert i kvoteOppsummeringUtils og brukteDagerUtils.
Eksporterer no frå @navikt/fp-uttaksplan.
