+++ packages/uttaksplan/index.ts
@@ -22,6 +22,6 @@ export { FjernAltIUttaksplanModal } from './src/FjernAltIUttaksplanModal';
 export { useErAntallDagerOvertrukketIUttaksplan } from './src/utils/kvoteOppsummeringUtils';
 export { UttaksperiodeValidatorer } from './src/utils/UttaksperiodeValidatorer';
 export { skalBesvareFlerbarnsdager } from './src/felles/LeggTilEllerEndrePeriodeFellesForm';
-export { sorterUttakPerioder, harPeriodeDerMorsAktivitetIkkeErValgt } from './src/utils/periodeUtils';
+export { sorterUttakPerioder, harPeriodeDerMorsAktivitetIkkeErValgt, getAntallUttaksdagerIVinduRundtFødsel } from './src/utils/periodeUtils';
 export { deltUttak } from './src/utils/forslag/deltUttak';
 export { ikkeDeltUttak } from './src/utils/forslag/ikkeDeltUttak';
